Seperate server and react functions
