+++ frontend/scripts/buildCache.ts
@@ -4,13 +4,16 @@
 import * as S3 from 'aws-sdk/clients/s3'
 import Axios from 'axios'
 import 'isomorphic-fetch'
+import { chain, keyBy } from 'lodash'
 import { seq } from 'promise-seq'
+import { buildCsvFromData } from '../src/components/DownloadAnalysisLink/BuildAnalysis'
 import { CONFIG } from '../src/config/config'
 import { DATASETS } from '../src/constants/datasets'
-import { Method } from '../src/constants/datatypes'
+import { Adequacies, AdequacyMode, Method } from '../src/constants/datatypes'
 import {
   getAdequacies, getCensusData, getRepresentativePoints, getStaticAdequacyUrl, getStaticDemographicsUrl, getStaticRPUrl
 } from '../src/services/api'
+import { getAdequacyMode } from '../src/utils/adequacy'
 import { safeDatasetHint } from '../src/utils/formatters'
 
 const METHODS: Method[] = ['driving_time', 'straight_line']
@@ -22,7 +25,7 @@ const METHODS: Method[] = ['driving_time', 'straight_line']
 */
 const s3Bucket = 'encompass-public-data'
 const s3 = new S3()
-const uploadToS3 = true
+const uploadToS3 = false
 
 main()
 
@@ -66,7 +69,21 @@ function cacheRPCensus() {
 }
 
 function cacheAdequacies() {
-  return seq(...DATASETS.map(dataset => async () =>
+  return seq(...DATASETS.map(dataset => async () => {
+    let representativePoints = await getRepresentativePoints({ service_area_ids: dataset.serviceAreaIds })
+    let census = await getCensusData({ service_area_ids: dataset.serviceAreaIds })
+    let storeLikeRps = representativePoints.map(_ => ({
+      ..._,
+      population: Number(_.population),
+      serviceAreaId: _.service_area_id,
+      demographics: census[_.service_area_id]
+    }))
+    if (uploadToS3) {
+      let rpParams = { Bucket: s3Bucket, Key: getS3Key(getStaticRPUrl(dataset)), Body: JSON.stringify(representativePoints), ContentType: 'application/json', ACL: 'public-read' }
+      let censusParams = { Bucket: s3Bucket, Key: getS3Key(getStaticDemographicsUrl(dataset)), Body: JSON.stringify(census), ContentType: 'application/json', ACL: 'public-read' }
+      s3.putObject(rpParams, s3Callback)
+      s3.putObject(censusParams, s3Callback)
+    }
     seq(...METHODS.map(method => async () => {
       console.log('  Getting Adequacy for ' + safeDatasetHint(dataset) + ' for ' + method)
       let adequacies = await getAdequacies({
@@ -75,12 +92,28 @@ function cacheAdequacies() {
         service_area_ids: dataset.serviceAreaIds,
         dataset_hint: safeDatasetHint(dataset)
       })
+
+      let hash = keyBy(storeLikeRps, 'id')
+      let storeLikeAdequacies: Adequacies = chain(storeLikeRps)
+        .map(_ => _.id)
+        .zipObject(adequacies)
+        .mapValues((_, key) => ({
+          adequacyMode: getAdequacyMode(
+            _, method, hash[key].serviceAreaId, dataset.serviceAreaIds
+          ),
+          id: _.id,
+          toClosestProvider: _.to_closest_provider,
+          closestProvider: dataset.providers[_.closest_providers[0]]
+        }))
+        .value()
+      let CSVResult = buildCsvFromData(
+        method, dataset.serviceAreaIds, storeLikeAdequacies, storeLikeRps, true)
       if (uploadToS3 && adequacies) {
         let adequacyParams = { Bucket: s3Bucket, Key: getS3Key(getStaticAdequacyUrl(dataset, method)), Body: JSON.stringify(adequacies), ContentType: 'application/json', ACL: 'public-read' }
         s3.putObject(adequacyParams, s3Callback)
       }
     }))
-  ))
+  }))
 }
 
 function getS3Key(s3URL: string) {
